feat: new variants per env form
