+++ frontend/src/component/feature/FeatureView/FeatureVariants/FeatureEnvironmentVariants/FeatureEnvironmentVariants.tsx
@@ -15,9 +15,8 @@ import {
     IFeatureVariant,
 } from 'interfaces/featureToggle';
 import { useMemo, useState } from 'react';
-import { EnvironmentVariantModal } from './EnvironmentVariantModal/EnvironmentVariantModal';
+import { EnvironmentVariantsModal } from './EnvironmentVariantsModal/EnvironmentVariantsModal';
 import { EnvironmentVariantsCard } from './EnvironmentVariantsCard/EnvironmentVariantsCard';
-import { VariantDeleteDialog } from './VariantDeleteDialog/VariantDeleteDialog';
 import useFeatureApi from 'hooks/api/actions/useFeatureApi/useFeatureApi';
 import { formatUnknownError } from 'utils/formatUnknownError';
 import useToast from 'hooks/useToast';
@@ -27,6 +26,8 @@ import { useChangeRequestApi } from 'hooks/api/actions/useChangeRequestApi/useCh
 import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
 import { useChangeRequestsEnabled } from 'hooks/useChangeRequestsEnabled';
 import { usePendingChangeRequests } from 'hooks/api/getters/usePendingChangeRequests/usePendingChangeRequests';
+import PermissionIconButton from 'component/common/PermissionIconButton/PermissionIconButton';
+import { Edit } from '@mui/icons-material';
 
 const StyledAlert = styled(Alert)(({ theme }) => ({
     marginBottom: theme.spacing(4),
@@ -62,9 +63,7 @@ export const FeatureEnvironmentVariants = () => {
     const [searchValue, setSearchValue] = useState('');
     const [selectedEnvironment, setSelectedEnvironment] =
         useState<IFeatureEnvironmentWithCrEnabled>();
-    const [selectedVariant, setSelectedVariant] = useState<IFeatureVariant>();
     const [modalOpen, setModalOpen] = useState(false);
-    const [deleteOpen, setDeleteOpen] = useState(false);
 
     const environments: IFeatureEnvironmentWithCrEnabled[] = useMemo(
         () =>
@@ -187,66 +186,20 @@ export const FeatureEnvironmentVariants = () => {
         }
     };
 
-    const addVariant = (environment: IFeatureEnvironmentWithCrEnabled) => {
+    const editVariants = (environment: IFeatureEnvironmentWithCrEnabled) => {
         setSelectedEnvironment(environment);
-        setSelectedVariant(undefined);
         setModalOpen(true);
     };
 
-    const editVariant = (
-        environment: IFeatureEnvironmentWithCrEnabled,
-        variant: IFeatureVariant
-    ) => {
-        setSelectedEnvironment(environment);
-        setSelectedVariant(variant);
-        setModalOpen(true);
-    };
-
-    const deleteVariant = (
-        environment: IFeatureEnvironmentWithCrEnabled,
-        variant: IFeatureVariant
-    ) => {
-        setSelectedEnvironment(environment);
-        setSelectedVariant(variant);
-        setDeleteOpen(true);
-    };
-
-    const onDeleteConfirm = async () => {
-        if (selectedEnvironment && selectedVariant) {
-            const variants = selectedEnvironment.variants ?? [];
-
-            const updatedVariants = variants.filter(
-                ({ name }) => name !== selectedVariant.name
-            );
-
-            try {
-                await updateVariants(selectedEnvironment, updatedVariants);
-                setDeleteOpen(false);
-                setToastData({
-                    title: selectedEnvironment.crEnabled
-                        ? 'Variant deletion added to draft'
-                        : 'Variant deleted successfully',
-                    type: 'success',
-                });
-            } catch (error: unknown) {
-                setToastApiError(formatUnknownError(error));
-            }
-        }
-    };
-
-    const onVariantConfirm = async (updatedVariants: IFeatureVariant[]) => {
+    const onVariantsConfirm = async (updatedVariants: IFeatureVariant[]) => {
         if (selectedEnvironment) {
             try {
                 await updateVariants(selectedEnvironment, updatedVariants);
                 setModalOpen(false);
                 setToastData({
                     title: selectedEnvironment.crEnabled
-                        ? `Variant ${
-                              selectedVariant ? 'changes' : ''
-                          } added to draft`
-                        : `Variant ${
-                              selectedVariant ? 'updated' : 'added'
-                          } successfully`,
+                        ? `Variants update added to draft`
+                        : 'Variants updated successfully',
                     type: 'success',
                 });
             } catch (error: unknown) {
@@ -266,23 +219,6 @@ export const FeatureEnvironmentVariants = () => {
                 title: toEnvironment.crEnabled
                     ? 'Variants copy added to draft'
                     : 'Variants copied successfully',
-                type: 'success',
-            });
-        } catch (error: unknown) {
-            setToastApiError(formatUnknownError(error));
-        }
-    };
-
-    const onUpdateStickiness = async (
-        environment: IFeatureEnvironmentWithCrEnabled,
-        updatedVariants: IFeatureVariant[]
-    ) => {
-        try {
-            await updateVariants(environment, updatedVariants);
-            setToastData({
-                title: environment.crEnabled
-                    ? 'Variant stickiness update added to draft'
-                    : 'Variant stickiness updated successfully',
                 type: 'success',
             });
         } catch (error: unknown) {
@@ -339,15 +275,6 @@ export const FeatureEnvironmentVariants = () => {
                         key={environment.name}
                         environment={environment}
                         searchValue={searchValue}
-                        onEditVariant={(variant: IFeatureVariant) =>
-                            editVariant(environment, variant)
-                        }
-                        onDeleteVariant={(variant: IFeatureVariant) =>
-                            deleteVariant(environment, variant)
-                        }
-                        onUpdateStickiness={(variants: IFeatureVariant[]) =>
-                            onUpdateStickiness(environment, variants)
-                        }
                     >
                         <StyledButtonContainer>
                             <PushVariantsButton
@@ -370,33 +297,51 @@ export const FeatureEnvironmentVariants = () => {
                                 onCopyVariantsFrom={onCopyVariantsFrom}
                                 otherEnvsWithVariants={otherEnvsWithVariants}
                             />
-                            <PermissionButton
-                                onClick={() => addVariant(environment)}
-                                variant="outlined"
-                                permission={UPDATE_FEATURE_ENVIRONMENT_VARIANTS}
-                                projectId={projectId}
-                                environmentId={environment.name}
-                            >
-                                Add variant
-                            </PermissionButton>
+                            <ConditionallyRender
+                                condition={Boolean(
+                                    environment.variants?.length
+                                )}
+                                show={
+                                    <PermissionIconButton
+                                        onClick={() =>
+                                            editVariants(environment)
+                                        }
+                                        permission={
+                                            UPDATE_FEATURE_ENVIRONMENT_VARIANTS
+                                        }
+                                        projectId={projectId}
+                                        environmentId={environment.name}
+                                    >
+                                        <Edit />
+                                    </PermissionIconButton>
+                                }
+                                elseShow={
+                                    <PermissionButton
+                                        onClick={() =>
+                                            editVariants(environment)
+                                        }
+                                        variant="outlined"
+                                        permission={
+                                            UPDATE_FEATURE_ENVIRONMENT_VARIANTS
+                                        }
+                                        projectId={projectId}
+                                        environmentId={environment.name}
+                                    >
+                                        Add variant
+                                    </PermissionButton>
+                                }
+                            />
                         </StyledButtonContainer>
                     </EnvironmentVariantsCard>
                 );
             })}
-            <EnvironmentVariantModal
+            <EnvironmentVariantsModal
                 environment={selectedEnvironment}
-                variant={selectedVariant}
                 open={modalOpen}
                 setOpen={setModalOpen}
                 getApiPayload={getApiPayload}
                 getCrPayload={getCrPayload}
-                onConfirm={onVariantConfirm}
-            />
-            <VariantDeleteDialog
-                variant={selectedVariant}
-                open={deleteOpen}
-                setOpen={setDeleteOpen}
-                onConfirm={onDeleteConfirm}
+                onConfirm={onVariantsConfirm}
             />
         </PageContent>
     );
